Adicionando API de funcionario
